requirements.txt
python-telegram-bot==20.7
python-dotenv==1.0.0
Pillow==10.0.1
pandas==2.1.1
openpyxl==3.1.2
# redis==5.0.1  # Render'da Redis yo'q
xlrd==2.0.1
